Interactive Items with Rayhits
Interactive Item script can be added onto any object within the game, as long as the object has a collision and the object tag is "Interactive" the ray-hit cast will interact with the object.

A door object will be given an interactive item script and within a serializefield string, you can add what you want the text to say with given object.